requirements.txt
streamlit==1.38.0
plotly==5.24.0
openai==1.45.0
fpdf2==2.7.9
python-dotenv==1.0.1
pandas==2.2.2
matplotlib
